DEV-151
Start implementing the size functions

+++ include/hil_rig_protocol/application/application_types.h
@@ -105,7 +105,7 @@ typedef struct
     const uint8_t* data;
 
     /** Number of readable bytes at data. */
-    size_t size;
+    uint32_t size;
 } HIL_Application_Byte_Span_T;
 
 /**
